+++ CHANGELOG.md
@@ -12,6 +12,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Support for `cursor` field on `v1/user/breached` of `user-intel` service.
 - Support for `severity` field on `v1/user/breached` and `v2/user/breached` of `user-intel` service.
 - `/v1/breach` endpoint support on `user-intel` service.
+- `vault_parameters` and `llm_request` fields support on Redact service.
 
 ## 4.1.1 - 2024-10-17